Add files via upload
Welcome to the ultimate time and weather control script for your test server! This script has been remastered to simplify the process of granting yourself permissions to change the time and weather. If you've ever struggled with setting permissions or had to use sv_lan 1, this is the solution you've been looking for.

Important: This script will grant all players the ability to change the time and weather on your server. Therefore, it is highly recommended to use this script only for your test server. For public or open servers, please refer to the original script that includes proper admin permissions.

If you encounter any issues or have questions, join my Discord and create a ticket: Discord Link:
https://discord.gg/99NcnNggcy

Original Script by Vespura:
Credit to Vespura for the original script, which includes admin permissions. You can find it here.
https://forum.cfx.re/t/vsync-v1-4-0-simple-weather-and-time-sync/49710


Enjoy full control over your test server's environment!